+++ releasenotes.md
@@ -1,9 +1,10 @@
 # `release notes`
 
-`next`
+`1.3.0`
 - #16 make CLI argument negatable
 - refactor: handling defaults
 - #19 added --batchsize CLI parameter
+- #23 added --listconfig CLI parameter
   
 `1.2.0`
 - #14 added --targetdir (-t) CLI parameter